Release Controller: Extend query params and remove releases routes in other controllers
