+++ server/src/artist/artist.controller.ts
@@ -33,6 +33,12 @@ class Selector extends IntersectionType(ArtistQueryParameters.SortingParameter) 
 
 	@IsOptional()
 	@ApiPropertyOptional({
+		description: 'If true, only artists that have at least one album will be returned'
+	})
+	albumArtistOnly?: boolean;
+
+	@IsOptional()
+	@ApiPropertyOptional({
 		description: 'Filter artists by genre'
 	})
 	@TransformIdentifier(GenreService)
